% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{gene_deviations}
\alias{gene_deviations}
\title{Compute deviations of gene expression for reference levels.}
\usage{
gene_deviations(expr, reference_levels)
}
\arguments{
\item{expr}{The expression matrix.}

\item{reference_levels}{A vector with a reference level for each gene.}
}
\value{
A table of deviations, with columns \code{sample} and
        \code{deviation}.
}
\description{
Compute deviations of gene expression for reference levels.
}
