% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{gene_medians}
\alias{gene_medians}
\title{Compute median expression levels for all genes.}
\usage{
gene_medians(expr)
}
\arguments{
\item{expr}{The expression matrix.}
}
\value{
A named vector with gene medians.
}
\description{
Compute median expression levels for all genes.
}
