% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease.R
\name{list_diseases}
\alias{list_diseases}
\title{Retrieve information about diseases.}
\usage{
list_diseases()
}
\value{
A data.frame containing disease abbreviations and descriptions.
}
\description{
Retrieve information about diseases.
}
\examples{
dplyr::filter(list_diseases(), abbreviation == "LUAD")

}
