% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{most_variable_profiles}
\alias{most_variable_profiles}
\title{Select the most abundant profiles from an expression matrix.}
\usage{
most_variable_profiles(expression, n = 100L)
}
\arguments{
\item{expression}{The expression matrix.}

\item{n}{The number of profiles to select (100, by default).}
}
\value{
A selection of rows from the input matrix.
}
\description{
Select the most abundant profiles from an expression matrix.
}
