% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{omicsdata}
\alias{omicsdata}
\alias{omicsdata-package}
\title{Package development tools for R.}
\description{
More details on the course at \url{https://sales.bio.unipd.it/teaching.html}
}
\author{
\strong{Maintainer}: Gabriele Sales \email{gabriele.sales@unipd.it} (\href{https://orcid.org/0000-0003-2078-5661}{ORCID})

}
