% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{sequencing_depth}
\alias{sequencing_depth}
\title{Compute the per-sample sequencing depth.}
\usage{
sequencing_depth(counts)
}
\arguments{
\item{counts}{The count matrix.}
}
\value{
An integer vector, holding the sequencing depth for each sample.
}
\description{
Compute the per-sample sequencing depth.
}
\examples{
library(airway)
data(airway)
sequencing_depth(assays(airway)$counts)

}
