% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{fetch_tcga_dataset}
\alias{fetch_tcga_dataset}
\title{Retrieve gene expression datasets for a disease.}
\usage{
fetch_tcga_dataset(disease)
}
\arguments{
\item{disease}{A disease code. See \code{\link{list_diseases}}.}
}
\description{
Retrieve gene expression datasets for a disease.
}
