% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{omicsdata}
\alias{omicsdata}
\alias{omicsdata-package}
\title{Support package for the "Omics Data" course.}
\description{
Collection of utility functions and data to be used in the "Omics Data" course.
}
\author{
\strong{Maintainer}: Gabriele Sales \email{gabriele.sales@unipd.it} (\href{https://orcid.org/0000-0003-2078-5661}{ORCID})

}
